refactor: remove enums module
